+++ express/apps.py
@@ -7,6 +7,7 @@ class ExpressConfig(AppConfig):
     name = 'express'
 
     def ready(self):
-    	self.module.autodiscover('services') # auto load services.py from each (installed) app
+    	# auto load apis indicated by services.py, models.py from each (installed) app
+    	self.module.autodiscover('services', 'models')
     	for url in self.module.services.urls:
     		logger.info('[express: url] ' + str(url.regex.pattern) + ' -- (name: ' + url.name + ')')