+++ testa/models.py
@@ -1,8 +1,10 @@
 from django.db import models
-from express.decorators import serve, serve_unprotected
+from express.decorators import serve, serve_unprotected, url
 
 # Create your models here.
 
+@url('/absolute/db/device')
+@url('db/device')
 @serve_unprotected
 class Device(models.Model):
 	"""docstring for Device"""